[NEW] use client.once for setupAlavirien
